requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
langchain==0.1.13
langchain-community==0.0.29
langchain-google-genai>=0.0.1,<0.1.0  
langchain-core==0.1.33
pypdf==4.1.0
pymilvus==2.3.7
qdrant-client==1.8.2
google-generativeai==0.3.2